k8s-microservices/requirements_api.txt
flask
mysql-connector-python
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-flask
coralogix-opentelemetry
opentelemetry-instrumentation-mysql
flask-cors
